refactor(p3): share fs/git adapter (makeFsGit) + unwrapGit via @kolu/terminal-dag
kolu-watcher (P3) must serve the remote host's fs/git from the SAME kolu-git
impl kolu-server uses locally — so factor LocalTerminalEndpoint's localFs/localGit
and the GitResult→ORPCError unwrap into @kolu/terminal-dag as makeFsGit(log) +
unwrapGit, consumed by both. local.ts now destructures makeFsGit(log) (behavior
identical — same kolu-git calls); server/unwrapGit.ts re-exports so router.ts's
import path is unchanged. typecheck + biome clean.
